use doubles instead of ints

+++ amatch/ext/vector.c
@@ -4,7 +4,7 @@ Vector *Vector_new(int len)
 {
     Vector *v = ALLOC(Vector);
     if (v == NULL) rb_raise(rb_eNoMemError, "couldn't malloc vector");
-    v->ptr = ALLOC_N(int, len + 1);
+    v->ptr = ALLOC_N(double, len + 1);
     if (v->ptr == NULL) rb_raise(rb_eNoMemError, "couldn't malloc vector data");
     v->len = len;
     return v;
@@ -13,7 +13,7 @@ Vector *Vector_new(int len)
 void vector_print(Vector *v)
 {
     int i;
-    for(i = 0; i < v->len; i++) printf("%d", v->ptr[i]);
+    for(i = 0; i < v->len; i++) printf("%lf", v->ptr[i]);
     puts("");
 }
 
@@ -23,11 +23,12 @@ void vector_destroy(Vector *v)
     xfree(v);
 }
 
-int vector_minimum(Vector *v) {
+double vector_minimum(Vector *v)
+{
     int i;
-    int min;
+    double min;
 
-    if (v->len == 0) return -1;
+    if (v->len == 0) return -1.0;
     min = v->ptr[0];
     for (i = 1; i <= v->len; i++) {
         if (min > v->ptr[i]) min = v->ptr[i];
@@ -35,7 +36,7 @@ int vector_minimum(Vector *v) {
     return min;
 }
 
-int vector_last(Vector *v) {
+double vector_last(Vector *v) {
     return v->ptr[v->len];
 }
   /* vim: set et cindent sw=4 ts=4: */
